api/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Vector database
qdrant-client==1.7.3

# OpenAI
openai==1.12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.26.0

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
tenacity==8.2.3
